+++ R/add_northarrow.R
@@ -43,7 +43,7 @@
 #' frames.b <- add_northarrow(frames, colour = "white", position = "bottomleft")
 #' frames.b[[100]]
 #' 
-#' @seealso \link{frames_spatial} \link{frames_graph} \link{animate_frames}
+#' @seealso \code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
 #' @export
 
 add_northarrow <- function(frames, height = 0.05, position = "bottomright", x = NULL, y = NULL, colour = "black", size = 1,