+++ R/add_timestamps.R
@@ -36,7 +36,7 @@
 #' frames.b <- add_timestamps(frames, m, type = "label")
 #' frames.b[[100]]
 #' 
-#' @seealso \link{frames_spatial} \link{frames_graph} \link{animate_frames}
+#' @seealso \code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{animate_frames}}
 #' @export
 
 add_timestamps <- function(frames, m, x = NULL, y = NULL, ..., verbose = TRUE){