+++ R/animate_frames.R
@@ -58,7 +58,7 @@
 #' # animate frames as mov
 #' animate_frames(frames, out_file = tempfile(fileext = ".gif"))
 #' }
-#' @seealso \link{frames_spatial} \link{frames_graph} \link{animate_frames}
+#' @seealso \code{\link{frames_spatial}} \code{\link{frames_graph}} \code{\link{join_frames}}
 #' 
 #' @export